Refactoring
* Changed names of model types and model parameters
  * Use longer more descriptive names
  * Use same length names for visibility
  * Collapse two models into one (tf_ising and tf_rf_nn_ising)
* Apply static cast where possible for correctness and to avoid implicit conversion warning
* Implementing resume from file still


Former-commit-id: c6925745169a0a6d18a7a713e4f5e81541c7d9bf

+++ source/math/class_eigsolver.h
@@ -256,7 +256,7 @@ void class_eigsolver::eig(const Eigen::EigenBase<Derived> &matrix,
                           const bool compute_eigvecs_,
                           const bool remove_phase_   )
 {
-    eig<type,form>(matrix.derived().data(),matrix.rows(), compute_eigvecs_,remove_phase_);
+    eig<type,form>(matrix.derived().data(), static_cast<int>(matrix.rows()), compute_eigvecs_,remove_phase_);
 }
